refactor package names and stucture
