--- components/org.wso2.carbon.identity.organization.management.enterprise.login.mgt.core/src/main/java/org/wso2/carbon/identity/organization/management/enterprise/login/mgt/core/model/Configuration.java
@@ -1,54 +1,0 @@
-/*
- * Copyright (c) 2022, WSO2 Inc. (http://www.wso2.com). All Rights Reserved.
- *
- * This software is the property of WSO2 Inc. and its suppliers, if any.
- * Dissemination of any information or reproduction of any material contained
- * herein in any form is strictly forbidden, unless permitted by WSO2 expressly.
- * You may not alter or remove any copyright or other notice from copies of this content.
- */
-
-package org.wso2.carbon.identity.organization.management.enterprise.login.mgt.core.model;
-
-import java.util.List;
-
-/**
- * Enterprise login management configuration object.
- */
-public class Configuration {
-
-    private List<String> services;
-    private List<String> emailDomains;
-    private String organization;
-
-    /**
-     * Constructor of configuration.
-     */
-    public Configuration(List<String> services, List<String> emailDomains) {
-        this.services = services;
-        this.emailDomains = emailDomains;
-    }
-
-    public List<String> getServices() {
-        return services;
-    }
-
-    public void setServices(List<String> services) {
-        this.services = services;
-    }
-
-    public List<String> getEmailDomains() {
-        return emailDomains;
-    }
-
-    public void setEmailDomains(List<String> emailDomains) {
-        this.emailDomains = emailDomains;
-    }
-
-    public String getOrganization() {
-        return organization;
-    }
-
-    public void setOrganization(String organization) {
-        this.organization = organization;
-    }
-}
